+++ matchms/filtering/__init__.py
@@ -108,6 +108,8 @@ from matchms.filtering.metadata_processing.make_charge_int import \
     make_charge_int
 from matchms.filtering.metadata_processing.repair_adduct_and_parent_mass_based_on_smiles import \
     repair_adduct_and_parent_mass_based_on_smiles
+from matchms.filtering.metadata_processing.repair_adduct_based_on_parent_mass import \
+    repair_adduct_based_on_parent_mass
 from matchms.filtering.metadata_processing.repair_inchi_inchikey_smiles import \
     repair_inchi_inchikey_smiles
 from matchms.filtering.metadata_processing.repair_not_matching_annotation import \
@@ -185,6 +187,7 @@ __all__ = [
     "remove_peaks_around_precursor_mz",
     "remove_peaks_outside_top_k",
     "repair_adduct_and_parent_mass_based_on_smiles",
+    "repair_adduct_based_on_parent_mass",
     "require_correct_ionmode",
     "repair_inchi_inchikey_smiles",
     "repair_parent_mass_is_molar_mass",