+++ matchms/filtering/default_pipelines.py
@@ -34,6 +34,7 @@ REQUIRE_COMPLETE_METADATA = [msfilters.require_precursor_mz,
 REPAIR_ANNOTATION = [
     (msfilters.repair_smiles_of_salts, {'mass_tolerance': 0.1}),
     (msfilters.repair_parent_mass_is_molar_mass, {'mass_tolerance': 0.1}),
+    (msfilters.repair_adduct_based_on_parent_mass, {'mass_tolerance': 0.1}),
     (msfilters.repair_adduct_and_parent_mass_based_on_smiles, {'mass_tolerance': 0.1}),
     msfilters.repair_not_matching_annotation,
     (msfilters.derive_annotation_from_compound_name, {"mass_tolerance": 0.1}),